+++ gdal/ogr/ogrsf_frmts/avc/avc_misc.c
@@ -239,7 +239,7 @@ GBool AVCFileExists(const char *pszPath, const char *pszName)
  * that list rather than all this stating that can be very expensive
  * in some circumstances.  However, at least with Carl's fix this is
  * somewhat faster.
- * see: http://buzilla.remotesensing.org/show_bug.cgi?id=314
+ * see: http://bugzilla.remotesensing.org/show_bug.cgi?id=314
  **********************************************************************/
 char *AVCAdjustCaseSensitiveFilename(char *pszFname)
 {
@@ -277,7 +277,7 @@ char *AVCAdjustCaseSensitiveFilename(char *pszFname)
     }
 
     pszTmpPath = CPLStrdup(pszFname);
-    nTotalLen = strlen(pszTmpPath);
+    nTotalLen = (int)strlen(pszTmpPath);
 
     /*-----------------------------------------------------------------
      * Try all lower case, check if the filename is OK as that.
@@ -437,7 +437,7 @@ int  AVCPrintRealValue(char *pszBuf, int nPrecision, AVCFileType eType,
 
         sprintf(szBuf, "%10.7E", 123.45);
         numExpDigits = 0;
-        for(i=strlen(szBuf)-1; i>0; i--)
+        for(i=(int)strlen(szBuf)-1; i>0; i--)
         {
             if (szBuf[i] == '+' || szBuf[i] == '-')
                 break;
@@ -489,7 +489,7 @@ int  AVCPrintRealValue(char *pszBuf, int nPrecision, AVCFileType eType,
     if (numExpDigits > 2)
     {
         int n;
-        n = strlen(pszBuf);
+        n = (int)strlen(pszBuf);
 
         pszBuf[n - numExpDigits]    = pszBuf[n-2];
         pszBuf[n - numExpDigits +1] = pszBuf[n-1];
