+++ gdal/ogr/ogrsf_frmts/gme/ogrgmejson.cpp
@@ -2,7 +2,7 @@
  * $Id$
  *
  * Project:  Google Maps Engine API Driver
- * Purpose:  GME GeoJSON helpper function Implementations.
+ * Purpose:  GME GeoJSON helper function Implementations.
  * Author:   Wolf Beregnheim <wolf+grass@bergenheim.net>
  *           (derived from Geo JSON driver by Mateusz)
  *
@@ -491,7 +491,7 @@ static int json_gme_double_to_string(json_object *pjo,
     }
     /* drop trailing zeroes */
     *(++p) = 0;
-    size = p-buf;
+    size = static_cast<int>(p-buf);
   }
   printbuf_memappend(pb, buf, size);
   return size;
